Value rewriting, part 3: Objects now hold a 'register' in them.
This is annoying in some cases because checking that an object has a
valid value is now something like

object->inner_value->flags & whatever

But this has some benefits. One is that values can be unboxed from
objects and then act like normal. One such case is in o_jump_if,
where the contents of an object are unboxed, making o_jump_if work
the same regardless of if something is in an object or not.

With this commit, all of values in Lily that were holding
collections of flags/value/sig are now using inner values. This has
brought about a lot of nice things (like generic_assignment!)

Now, all that is left is to do some naming changes, and make things
take values instead of sig+value pairs and that sort stuff.