Value rewriting, part 4: Support for proper vals, some naming changes.
lily_value -> lily_raw_value
    A raw value is a value that doesn't have flags or a signature to go with
    it. These are no longer common, which is rather nice.

lily_vm_register -> lily_value
    Now, lily_value corresponds to a 'proper' value. A proper value is one
    that has a signature, flags, and a raw inner value.

Lots of functions were rewritten so they take a proper value instead of a
sig + raw value combo.

The gc still keeps a sig and raw value combo for gc_entry. This is because
there is little value in creating a proper value. For one, it would mean an
extra allocation. It can't save the proper value, because the proper value
may be a register that will change the sig+value (leaving the gc very, very
confused).

Added lily_deref_unknown_raw_val for destroying raw values.

This rewrite was necessary for a few reasons:
* Subscript and subscript assign of lists+hashes was atrocious. There were
  many, many bugs over time from forgotten parts here and there.
* It allowed making generic_assignment, which is a magic function that copies
  data from one proper value to another. This was used to make subscript and
  subscript assign saner and also more readable.
* This restricts ref/deref stuff in vm to fewer places, meaning fewer
  potential bugs.
* Proper values can be passed to more things, like deref and vm calls,
  resulting in fewer arguments to calls.
* I have a feeling that this will make foreign functions much, much easier.
  A list append would need to just expand list elements and call for
  generic_assignment without having to worry about type stuff.