The vm now marks literals loaded with VAL_IS_PROTECTED.
VAL_IS_PROTECTED is a new flag for values. This flag indicates that the
associated value should never receive a ref or a deref. This is nice because
it saves from doing ref/deref to literals that will never be deleted until the
very end. The overhead is very small (where things used to check for nil, they
now check for nil|protected).