sanity.ly reorganized a bit.
It now uses a list of methods to call each method so that list's size member
is also demonstrated. Yay!

A few silly tests were removed (like testing addition). These were relics from
when the language was a lot less stable.

Also went to basic assignments and added in tests for types that have been
added since the basic assignment test was created. Lots more comments for the
first tests too.

Overall, not much change in line count. However, there are fewer tests since
a couple were merged together (while and for..in tests merged) too.

Finally, and most importantly, a test for object defaulting has been added,
because I believe there's been issues with that in the past.