All blocks can be single-line now. Also, oops fix for do...while.
do...while has been broken since first introduced: It was jumping back to the
top on the opposite condition. This was never found because there's no test
for it.

Merged multi-line block handling into a function, and that's now used by the
main non-if blocks (do...while, while, and for..in) to handle the multi-line
case. The single-line case is basic though (just statement(parser, 0)).

I also reworked some of the handler comments in parser, one of which was
rather redundant.