Parser: expression no longer takes flags. Emitter: Clear ast after eval.
The parser was adjusted a bit so that it no longer needs EX_SAVE_AST or
EX_NEED_VALUE. This was done by making a declaration list's = push the var and
the assign op, since that was the only time a value wasn't needed.

EX_SAVE_AST was removed by making expression no longer auto-emit, and instead
leave that job to the caller. All blocks needed to save the ast. The only two
cases where it didn't need to be saved were standard expressions via
statement, and the declaration list's =.

I fixed all of the emitter's functions that evaluate asts to have
lily_emit_eval_* as their name, then clear the ast pool for the next
expression.