C's argv now wrapped as a global argv in Lily.
This is a step toward wrapping it under a package. Used blastmaster to ensure
that this did not introduce a memory leak.

Also, added lily_pkg_sys which will eventually add argv as part of a new sys
package. But, for now, it just adds argv to the list of globals.