Added tree_oo_call, which handles abc.xyz type calls...
This makes it so emitter, not parser, is responsible for doing class callable
lookups. Since emitter does evaluating of asts already, it's much, much easier
to have it eval and then determine the resulting class to do a call against
instead of parser doing it.

The ast now has a growable string 'pool' that it shares with emitter. Parser
copies strings into it, and tree_oo_call has indexes to the strings inside,
allowing ast to grow it without fear of corrupting anything when realloc-ing.

An axe has been taken to determine_ast_sig in parser. Nothing remains. Good
riddance.