Redo tree_typecast to be entered, fixing GH #21.
@(object: 1 + 1) resulting in 'cannot do object + integer' was the result of
a binary tree taking over the typecast tree as a left side. This is fixed by
entering a typecast tree upon seeing the '@(' token. This means that binary is
unable to mess with the typecast tree (this is how subscripts, list building,
and most other trees work). I've added a test for this behavior to sanity.ly,
though I doubt it will ever occur again.