ast: Unary uses absorb merge for everything. Fixes GH #20.
The problem was that unary wasn't merging in call trees and other uncommon
kinds of trees. Now, an absorbing merge is used to swallow everything, solving
the problem.

Also, fixed merge_absorb to not set ap->active in the event that the active
sent is not actually ap->active.

Test added to sanity.ly to prevent this from happening again.