BIG CHANGE: Shove everything into a src/ dir, use autotools to compile now.
This was done because I intend to start an apache module which I believe will
need autotools to put together. This also streamlines the build process by
getting rid of a makefile that wasn't being updated properly at all with
regard to dependencies.

I've also added a couple of new Makefile targets: One to run the sanity test,
and another to run the blastmaster suite of tests.