Added the beginning of an apache module. Yay!
Added a rule in configure.ac to build the apache mod only on demand. It's off
by default in case people don't have the headers to build it but want to build
the standard lily.

The apache mod has a Makefile.am that replicates most of what's in the
Makefile generated by `apxs -g -n "lily"' (as far as apache commands from
within it). This will make it easy to test the apache mod as it's developed.

Note: The apache module does not do anything except to show that it's excited
that it works when polled. The impl's for printing are stubbed out, and the
mod doesn't create a working parser.

It's a start!