The interpreter now takes runner data in a void *.
lily_impl_puts now takes a void * as the first argument. This is the same data
that runners pass as the new first argument to lily_new_parse_state. This is
passed around to lexer and vm (with vm handing it to debug). The interpreter
does not touch this pointer at all: It's strictly for the runner.

This was done so that the apache module can obtain the request_rec without
having to resort to global variables or other evilness.