lexer mode is now independent of file type, and all loads now need it.
This was done to reflect the fact that a particular lexer input type is not
going to always be tagged/untagged. A file may want to be tagged or untagged.
A string may want to be tagged. So all of the parser's load_* calls take the
mode now.

Also, lily_lexer_mode -> lily_lex_mode, to be consistent with lily_lex_state
(both are lily_lex_* after the change). The modes are now lm_tags and
lm_no_tags to better reflect what the modes do.

I also changed lily_load_special to take the mode before the source. This
makes it consistent with the other lily_load_* calls that take the mode before
the particular source.

All of the runners have been updated to now pass the mode they want, except
for apache which had the args reordered.