Hash entries now deref keys, hash elems now copy key flags.
I found a memory leak while doing apache (again). This time, if a hash has a
str key that has one ref and isn't protected, the str will never get free'd.
So the hash derefs the keys is gets. This seems right, since I think hashes
ref their keys as well.

Additionally, hash element creation copies over keys. This was necessary to
keep the hash elem deref from thinking that a literal key was not a literal
key (and trying to deref it).