Postfix typecasts: `value.@(type)`. Fixes GH #30.
Removed the old syntax and hopefully all of the old references to it.
Also, all of the stuff near expression_oo in expression/expression_value was
merged into it for simplicity.

Updated sanity.ly to remove all old typecasts in favor of new ones. I like the
new syntax already.