apache: Fix Makefile.am. Also, bind GET vars now. Fixes #35.
The Makefile.am forgot to add lily_class_funcs.c, which was causing errors
when trying to load the lily module.

This makes GET vars available as server::get::*, but assigning to them does
not work yet (I'll fix that soon). The server will not bind GET vars with
invalid names (they wouldn't be available anyway!).