apache: Merge env and get scanning code. This makes server::get a hash.
I feel a bit dumb now. So...I started thinking about server::post, and
realized that conditional vars wouldn't work. If the server isn't doing a post
request, then attempts to access post variables would cause the parser to
raise messages about being unable to find stuff.

So I decided to make server::get a 'hash[str, str]' type. This is suitable for
now, and may be updated in the future to something fancier for more fancy
requests. I'm intending to do server::post this way to, and for the same
reason.

Since server::get and server::env are pretty similar, the code concerning
both of them was merged together, resulting in some code cutting.