list::apply, finally. Foreign calls now possible! Fixes #31.
This adds a few functions that allow functions to call back into the vm. At
this point, the only function that calls the vm looks like this:

list::apply(list[T], method(T):T):nil

where T is the type of the list. This calls the given method on each element
of the list. The return value of the method is assigned to the result pumped
out by the callee method.

Additionally, functions can now be added to the method stack so that a
function that fails shows up properly in the error stack. All the runners have
been updated to show stack trace properly.