Regression fix: Adding list::apply broke list::append.
The argument checking in emitter should always grab 'self' from the first
argument given, regardless. The reason that list::append didn't break was that
the argument it gets is a complex one (a method) and so the template accesses
had something to get.

However, list::append with list[integer] meant that it took 'integer', which
had no inner types to do a template access against. So...crash.

Added a test to sanity.ly to make sure the list::append oops doesn't happen
again.