Added hash::get, call returns now handle simple templates.
hash::get is the same thing as a subscript, except it takes a param to act as
a default value instead of raising an error for a key not being found. I made
try_lookup_hash_elem into lily_try_lookup_hash_elem (it's now exported) so
that hash::get wouldn't have to duplicate code.

Also, I fixed call returns to grab the appropriate template if the return is
a template (since hash::get returns the value type of the hash given). This
works now, but eventually a full template replacement will be necessary.

This call was made for apache, since trying to grab hash elements that aren't
there fails otherwise.