Print class names for functions in emitter and stack trace. Fixes #42.
This means that calling append with invalid args will now show it as
list::append instead of just append. Also, the stack trace now shows that sort
of stuff too.

This way of doing it may need to be changed in the future, since storing the
names only allows for a single level of class depth (so nested classes will
need to be figured out later).

Also, updated all of the runners wrt how they print out their stack traces.
I didn't touch the apache module because that doesn't dump stack just yet.