vm: Slash a bunch of code from the o_build_list implementation.
This will make static list building a hair slower...but it allows cleaving
away a large block of code, as well as ensuring that aren't likely to be any
bugs in it. Considering the pain that I got from trying to fast path
subscript/subscript assign in the past, I'd MUCH rather go for this way.