lily_build_ensure_sig, build non-templated sigs for call returns.
There is a new function called lily_build_ensure_sig. This function takes
the symtab, a lily class, a siglist count, a siglist, and a siglist offset.
From these, it builds a sig and uses the siglist (without mallocing) to do a
lookup. If the sig doesn't exist, then it's created and installed in the
symtab.

This function is used in two ways: The emitter uses it to build a signature
that doesn't have templates from one that does (this is necessary for returns
like list[A] or maybe[A]). The parser now uses this to build signatures for
simple templated types.

This will cut down on memory usage from the parser (no more creating a list
of integer only to have to throw it away). I also unified parsing for lists,
hashes, and tuples to cut down on redundancy.

I also kicked out the circularity + template counting code into a new symtab
function called finalize_sig. Symtab is a bit rougher until finalize sig is
finally gone.