Merge most fail tests into pre-commit-hook.py. Also a symtab fix.
I accidentally left some fprintfs in the code, so those were removed. More
importantly, I removed all of the failing tests that I determined to be too
simple to be in a file and merged them into pre-commit-hook.py. I've decided
to leave a few around because they're more complicated than a couple lines.

This is good because it means that fewer tests will simply rot or never be
used as often as they should.