Exceptions, round 3: Handle other exceptions now.
All errors that can be raised at vm time are now initialized in classes, and
handled during error catching. The exception test now checks for RangeError
which is raised when a subscript is out of range.

Also fix a wildly incorrect comment over symtab's lily_class_by_name.

Did an adjustment in parser so that blank except entries work in multi-line
blocks.