Exceptions, round 4: Add parents to classes, make lily_seed_symtab more sane.
For the first one, the parent of a class is given as a string. This is
important, because future classes will need to do this. For example, a mysql
function needs to a particular class, it has no guarantee of what ID it will
have. So it will need to use a string for identity. Or...something.

For the second, class init now has a second pass where parents are
initialized to a non-NULL value. This is so that lily_class_by_id is not
called when all classes have not been filled in (and where there may be empty
ones!)

This is a step toward checking that 'except' gets something that can, you
know, actually BE raised. Also, inheritance!