Exceptions, round 5: Inheritance checks!
'except' now verifies that the given class derives from Exception, which is
pretty important. Additionally, the vm checks for inheritance when processing
exceptions. This means that if the 'except' class given is more general than
the raised exception, it works anyway. This pattern is used sometimes to catch
any exception and log it.

Also, fix classes to actually init parents (the line that set classes pos just
needed to be moved up. Small oopsie there).