Exceptions, round 6: Pop a try at the end of it.
If a 'try' block does not raise an exception, then the catch entry added needs
to be removed. If it is not, the vm will try to use it on the next exception
(with a near-certain crash).

This adds an instruction called o_pop_try that just pops a try entry when
there is no exception raised.