Exceptions, round 7: Unwind stack when catching exceptions. New tests.
This one took a bit of time to figure out. However, it seems to work correctly
except that it doesn't reclaim the registers (it also needs to adjust the
current number of registers used). I'll fix that later though.

Added two tests: One catches an exception raised in a different function, the
other does a nested exception test.