Exceptions, round 8: Create a proper exception object, show it.
The parser now understands that adding 'as <name>' to the end of an 'except'
means that an exception object should be made and put into that var. So, for
example, 'except ValueError as V' would put the contents of the exception into
var V.

The vm will only create an exception object if 'as <name>' is specified. This
is an optimization: In cases where the user -just- wants to know that a
particular exception has been raised, this avoids creating a
somewhat-expensive exception object that will just get ignored.

The exception object created is fairly useless...but it can be passed to show
to prove that the contents are correct.

If the vm is creating a traceback object and runs out of memory, the following
NoMemoryError that gets raised is uncatchable. This is done intentionally to
keep the interpreter from doing an infinite loop trying to build traceback if
it's out of memory (because aft WILL test that).

Added a test that checks the new 'as <name>' syntax and shows the created
type.