Exceptions: Pop try blocks for continue, break, return.
This fixes a previous issue where one of the above inside a try block would
cause the block to never be unregistered by the vm. So the next catch could
possibly jump back into a block that the vm is no longer within.

Added BLOCK_TRY_EXCEPT to emitter. When the emitter reaches an except block,
BLOCK_TRY becomes BLOCK_TRY_EXCEPT. This also helps the emitter determine how
many try blocks that need to be popped (if any).