Emitter: New macro called lily_raise_adjusted.
This new macro combines writing out an adjust to the raiser and then raising
an error. The purpose of this is to reduce redundant code (shortening things
to one line or down by a line in many cases). Also, it makes things clearer
as to "this is doing an adjust" versus "this is not adjusting".