Emitter: BLOCK_* defines are numbers, not flags.
A long time ago, I thought that these might be useful as flags. However, that
time never came. These are not flags, so define them as numbers so that there
will be no confusion in the future.

IS_LOOP_BLOCK was changed to write out a separate check for each line, but
that's a small price to pay.