Emitter: Merge lily_emit_return + lily_emit_return_noval, fix an old bug.
This change means that the parser can parse a return value if needed, and give
the same function a valid ast, or NULL.

Additionally, a bug was discovered where the parser would crash when return
expected a value, but was given a function that did not return a value. This
bug has been around for...longer than I want to think about. I added a test to
pre-commit-hook.py to make sure the bug does not come back (though I doubt
that it will).