Emitter: eval_enforce_value. The end of 'ast->result == NULL' checks.
I noticed a pattern where lots of emit stuff was checking that the ast that it
evaluated returned a value. An ast can potentially not return a value if the
top of it is a function that doesn't return anything. print and printfmt are
both cases of this.

The emitter's call eval DOES check that a function returns a value if
necessary if that function is in another tree. However, with things like 'if',
'show', 'raise', etc., there's no 'if'/'show'/etc. tree to wrap around the
call.

Anyway, I wish I had thought of this a long time ago, because "crash when this
doesn't get a value" has been an issue more than once.