Symtab+Parser: Split get_intnum_literal...and fix it. :(
The idea behind grabbing for a literal with a value is that the symtab is
supposed to see if it has a literal with that value. If it does, the literal
that already has the value is returned. Instead of, you know, just blindly
making another literal.

Well...it turns out that at some random point, lily_get_intnum_literal had a
very important bit missing. A 'ret = lit;' was needed before the 'break'. But
since that wasn't there...integer and double literals were not unique.

Sanity.ly beforehand:
	3,930 allocs with 432,309 bytes used.
After:
	3,365 allocs with 409,709 bytes used.

-565 allocs, -22,600 bytes used.

Additionally, the broken up functions don't require a lily_raw_value, or a
class. This allowed some code to be removed from parser, which is nice.