Symtab: lily_ensure_unique_sig now takes flags, shuffled count to the back.
Flags comes after the class, and the count is now at the end. This makes it
so the last two parameters act as a range of sorts (start, then how many from
there). So it seems a bit sane.

Anyway, flags was added so that vararg sigs work. Otherwise,
    function b(integer a...) { ... }
is the same as
    function c(integer a) { ... }
which is not the case.

Emitter and parser have been updated appropriately.