Parser: Unify inner type collection for all types.
There used to be a separate loop for lists/hashes/tuples versus functions.
It's about the same size with the two loops unified (because of all the state
stuff. However, type collection for functions now uses lily_build_ensure_sig.
Because it uses this function, duplicate siglists are no longer malloced just
to be thrown away. Similarly, the () case was pretty wasteful too.

There's a very, very minor change to sanity.ly's memory usage.

This is just another part of reorganizing the internals of the interpreter.