Parser: A twofer: Cut down if/elif/else, and fix an unknown bug.
The bug was this:

if 1: elif 1: 1

This was because of if dispatching to elif/else directly but only once. The
bug was that the interpreter would report an extra '}' (because the if block
was being popped twice).

Pop 1: if -> elif (last expression, so pop).
Pop 2: if (returned to, so pop).

I fixed this by making if_handler dispatch to elif/else in a loop, and by
making elif_handler and else_handler leave block state alone. For extra
measure, I condensed all of if_handlers block pops to a pop at the end.

This adds a rather strange feature: In a single-line block, every condition
but the last one can be empty. I fully admit that this seems strange and may
not be wanted in the future.

This can be removed in the future, however. To do so, if_handler would need to
check emitter's expr_num before and after each single-line block. A
single-line block should have two expressions per elif: one for the condition,
and one for the statement under it. If you go this route, have things like
break, continue, (and return if it doesn't get an expression) bump up the
expr_num artificially. Otherwise 'if 1: break' will be wrongfully trapped.