Parser: For 'type x = value', push x to the ast properly.
Previously, the parser was using 'lily_push_push_sym' regardless of the depth
of the var. This was technically correct...except that the ast marks it as
a var (global) instead of a local var.

There is a difference to this. An assignment to a global is unlikely to be
optimized by the emitter's optimizing function. The optimization is a small
one (the last instruction is rewritten to write to a local, instead of it
emitting to a storage and then writing to a local).