Parser: Use an integer sig for new functions by default.
Previously, the parser was calling lily_try_sig_for_class, resulting in one
new empty function signature for each function declaration. Not only was it
made, but because ensure was never done on it, there were -multiple- blank
function signatures made.

Now it uses the signature of integer (of which there is only one), and swaps
the signature of the var created if it needs to do that. The result is 86
fewer allocations within sanity.ly, which is nice.