lily_ast_enter_tree -finally- takes 2 arguments.
The third argument, an initial value, was initially done as a hack by the old
expression engine to handle function calls. Once expressions were rewritten,
only static accesses of functions used this. So I changed that.

The third argument was NULL 99% of the time, so this is a good thing. :)