ast: Hey unary, use merge_value.
It is not possible to merge unary and have a value as a current, so the checks
being done were all impossible. If this does happen, then parser has some
explaining to do (because it's merging unary like it's a binary).

This started with registers being added (probably an accident) and was then
added to over time as I thought the previous code was sane. Well...it wasn't.