ast: Cleanup the header a bit and drop ast->arg_top.
It's much easier to simply loop to find arg_top and set it since nothing else
uses arg_top outside ast. There are definitely some more savings in ast field
removal but...

It's a waste of time. Removing that one field saves around 200 bytes of
memory, which is laughable because the current sanity.ly shows a use of nearly
400K for me.