vm: Stop re-implementing assignments. Use lily_assign_value!
To be fair, some of this code came from before there was a lily_assign_value.
However, now it's a lot of duplicated stuff. This sort of thing makes it
easy to have a hidden bug somewhere. o_return_val actually DID have a bug in
it (it was not properly handling returning to any).

But this is more about removing things that are just plain redundant.