Merge common binding in mod_lily and vm into lily_bind.{c,h}
I should have done this when I was making exceptions, instead of copy+pasting
big chunks of code from mod_lily.c. I also changed the functions that were
put in common space to take a symtab (no more making a signature to pass it!)
This means that each bind function has to do a symtab lookup, but that's okay.