symtab: Cleanup and reorganization. Finally fixes #7.
The main, important modules are in much better shape than when I first
started. Also, I'm much more aware of the parts that need to be redone/are
terrible.

I'm skipping over debug, even though I'm well aware that it's not all that
great. The reason is that I intend to change show into a function, as well as
break the code dumper into individual functions (the whole switchtable thing
is getting really, really nasty).

Time to add new things again! :)